Correctly handle animation ticks
Regeneration lengths can be different depending on the type, do this has to be a progress-percentage based system.
I also tweaked the fiery animation. Imho it's more "realistic", the cones start growing later on and they finish earlier.
The math might look like black magic but I'll create an explanation page for it, it's really not that hard.

@Suffril, if you don't like the timings in the animation you can tweak the x0 and x1 variables (TypeFiery:127). In short x0 is the percentage at which the cones start to grow, and x1 is the percentage at which the cones are at max size.