@Suffril I need help
There's something weird going on with synchronisation. In `onPlayerLoggedIn` we immediately synchronise to make sure the client has the current state and stuff that was saved in the world. It looks like it works fine at first, but it actually *only* works **the first time you enter a world**. If you leave and log back in it doesn't work. Restart minecraft: it's working again. Using some ~~System.out.println~~ debugger output I have discovered that the network handler for synchronization is never even called.

I have no idea wtf is going on here, could you take a look?