@Suffril I need help part 2
So I did some more investigating, and I *think* I've found a clue. In the fromBytes of the synchronization packet we check for a null player. Strangely enough this player *is* null when we first load, but not the first time we log into our world.
I'm really confused now. I also don't understand why this would cause the handler method to not even be called.